CrashCourse_OpenAISDK_Panaversity/Class04_OpenAISDK_CrashCourse/multi_agents/pyproject.toml
[project]
name = "multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.11",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
